Rename Contollers/ to Controllers/ to fix typo
Move all input handling files (MouseController, IMouse, IKeyboard,
IController, IGameInputHandler, KeyboardController) to the correctly
spelled Controllers/ directory.

Made-with: Cursor
