--- Contollers/IController.cs
@@ -1,4 +1,0 @@
-public interface IController
-{
-    void Update();
-}
