--- Contollers/IKeyboard.cs
@@ -1,6 +1,0 @@
-using Microsoft.Xna.Framework.Input;
-
-public interface IKeyboard : IController
-{
-    KeyboardState GetState();
-}
