--- Contollers/IMouse.cs
@@ -1,6 +1,0 @@
-using Microsoft.Xna.Framework.Input;
-
-public interface IMouse : IController
-{
-    MouseState GetState();
-}
