+++ Controllers/IController.cs
@@ -1,0 +1,4 @@
+public interface IController
+{
+    void Update();
+}
