+++ Controllers/IKeyboard.cs
@@ -1,0 +1,6 @@
+using Microsoft.Xna.Framework.Input;
+
+public interface IKeyboard : IController
+{
+    KeyboardState GetState();
+}
