+++ Controllers/IMouse.cs
@@ -1,0 +1,6 @@
+using Microsoft.Xna.Framework.Input;
+
+public interface IMouse : IController
+{
+    MouseState GetState();
+}
